🔧 chore[rust]: clear strict clippy warnings

+++ src-tauri/src/ocr/ppocr.rs
@@ -134,25 +134,17 @@ impl OcrEngine for PpOcrEngine {
             let left = bbox
                 .points
                 .iter()
-                .map(|p| p.x as f32)
+                .map(|p| p.x)
                 .fold(f32::MAX, f32::min)
                 .max(0.0) as u32;
             let top = bbox
                 .points
                 .iter()
-                .map(|p| p.y as f32)
+                .map(|p| p.y)
                 .fold(f32::MAX, f32::min)
                 .max(0.0) as u32;
-            let right = bbox
-                .points
-                .iter()
-                .map(|p| p.x as f32)
-                .fold(f32::MIN, f32::max);
-            let bottom = bbox
-                .points
-                .iter()
-                .map(|p| p.y as f32)
-                .fold(f32::MIN, f32::max);
+            let right = bbox.points.iter().map(|p| p.x).fold(f32::MIN, f32::max);
+            let bottom = bbox.points.iter().map(|p| p.y).fold(f32::MIN, f32::max);
             let width = (right - left as f32).round().max(1.0) as u32;
             let height = (bottom - top as f32).round().max(1.0) as u32;
